Define additional NFC pins
With the new board layout the NFC pins are actually connected to the olimex,
however the PN532 chip can miss-behave if the outputs are not properly set.

WIth this patch we define the IRQ and Hardware Request pins of the PN532
so that we can put them in a defined state.

EM-1748

Signed-off-by: Marco Teunissen van Manen <m.teunissenvanmanen@ultimaker.com>
Signed-off-by: Olliver Schinagl <o.schinagl@ultimaker.com>